requirements.txt
opencv-python
ffmpeg-python
torch
transformers
Pillow
pyttsx3
tk
